Move Collector to its own package
It's not common to import from main packages in Go. Move the Collector to its own package to have a standalone entry point.
